+++ tests/test_extensions/test_blocks/test_general_blocks.py
@@ -2,8 +2,7 @@
 from ... import util
 import unittest
 from pymdownx.blocks import block
-# import xml.etree.ElementTree as etree
-# import markdown
+import markdown
 
 
 class TestTypeFunctions(unittest.TestCase):
@@ -118,462 +117,188 @@ class TestTypeFunctions(unittest.TestCase):
         self.assertEqual(['this', 'that'], block.type_html_classes('this that'))
 
 
-# class TestGeneral(unittest.TestCase):
-#     """Test general cases."""
-
-#     def test_attribute_override(self):
-#         """Test that attributes cannot be overridden."""
-
-#         class AttrOverride(block.Block):
-#             NAME = 'override'
-
-#             OPTIONS = {
-#                 'attrs': [False, block.type_boolean],
-#             }
-
-#             def on_create(self, parent):
-#                 """Create."""
-
-#                 return etree.SubElement(parent, 'div')
-
-#         with self.assertRaises(ValueError):
-#             markdown.markdown(
-#                 '/// override\n///',
-#                 extensions=['pymdownx.blocks'],
-#                 extension_configs={'pymdownx.blocks': {'blocks': {AttrOverride: None}}}
-#             )
-
-#     def test_duplicate_blocks(self):
-#         """Test duplicate blocks."""
-
-#         with self.assertRaises(ValueError):
-#             markdown.markdown(
-#                 '/// override\n///',
-#                 extensions=['pymdownx.blocks'],
-#                 extension_configs={
-#                     'pymdownx.blocks': {
-#                         'blocks': {'pymdownx.blocks.admonition:Admonition': {'types': ['danger', 'danger']}}
-#                     }
-#                 }
-#             )
-
-
-# class TestBlockUndefinedOption(util.MdCase):
-#     """Test Blocks with undefined options."""
-
-#     class UndefinedBlock(block.Block):
-#         """Undefined option block."""
-
-#         NAME = 'undefined'
-
-#         def on_create(self, parent):
-#             """Create."""
-
-#             return etree.SubElement(parent, 'div')
-
-#     extension = ['pymdownx.blocks']
-#     extension_configs = {'pymdownx.blocks': {'blocks': {UndefinedBlock: None}}}
-
-#     def test_undefined_option(self):
-#         """An undefined option will cause the block parsing to fail."""
-
-#         self.check_markdown(
-#             R'''
-#             /// undefined
-#                 option: whatever
-
-#             content
-#             ///
-#             ''',
-#             '''
-#             <p>/// undefined
-#                 option: whatever</p>
-#             <p>content
-#             ///</p>
-#             ''',
-#             True
-#         )
-
-
-# class TestMiscCases(util.MdCase):
-#     """Test some miscellaneous cases."""
-
-#     class MiscBlock(block.Block):
-#         """A miscellaneous block."""
-
-#         NAME = 'misc'
-#         OPTIONS = {'test': ['whatever', block.type_string]}
-
-#         def on_create(self, parent):
-#             """Create."""
-
-#             return etree.SubElement(parent, 'div', {'test': self.options['test']})
-
-#     class MiscInline(block.Block):
-#         """A miscellaneous block."""
-
-#         NAME = 'miscinline'
-
-#         def on_create(self, parent):
-#             """Create."""
-
-#             return etree.SubElement(parent, 'span')
-
-#     extension = ['pymdownx.blocks', 'pymdownx.superfences']
-#     extension_configs = {'pymdownx.blocks': {'blocks': {MiscBlock: None, MiscInline: None}}}
-
-#     def test_general_config(self):
-#         """Test that content block should have a blank line between config."""
-
-#         self.check_markdown(
-#             R'''
-#             /// misc
-#                 test: misc
-#             content
-#             ///
-#             ''',
-#             '''
-#             <div test="misc">
-#             <p>content</p>
-#             </div>
-#             ''',
-#             True
-#         )
-
-#     def test_no_content(self):
-#         """Test config and no content."""
-
-#         self.check_markdown(
-#             R'''
-#             /// misc
-#                 test: misc
-#             ///
-#             ''',
-#             '''
-#             <div test="misc"></div>
-#             ''',
-#             True
-#         )
-
-#     def test_unfenced_config_and_no_content(self):
-#         """Test no fenced config and no content."""
-
-#         self.check_markdown(
-#             R'''
-#             /// misc
-#                 test: misc
-#             ///
-#             ''',
-#             '''
-#             <div test="misc"></div>
-#             ''',
-#             True
-#         )
-
-#     def test_content_after_only_config_block(self):
-#         """Test content after only config block."""
-
-#         self.check_markdown(
-#             R'''
-#             /// misc
-#                 test: misc
-#             ///
-#             more
-#             ''',
-#             '''
-#             <div test="misc"></div>
-#             <p>more</p>
-#             ''',
-#             True
-#         )
-
-#     def test_superfence_block(self):
-#         """Test blocks with fenced code content."""
-
-#         self.check_markdown(
-#             R'''
-#             /// misc
-#             ```python
-#             import foo
-#             ```
-#             ///
-#             ''',
-#             '''
-#             <div test="whatever">
-#             <div class="highlight"><pre><span></span><code><span class="kn">import</span> <span class="nn">foo</span>
-#             </code></pre></div>
-#             </div>
-#             ''',
-#             True
-#         )
-
-#     def test_superfence_inline(self):
-#         """Test blocks with fenced code content."""
-
-#         self.check_markdown(
-#             R'''
-#             /// miscinline
-#             ```python
-#             import foo
-#             ```
-
-#             Other content
-#             ///
-#             ''',
-#             '''
-#             <span><code>python
-#             import foo</code>
-
-#             Other content</span>
-#             ''',
-#             True
-#         )
-
-
-# class TestMiscYAMLFenceCases(util.MdCase):
-#     """Test some miscellaneous cases."""
-
-#     class MiscBlock(block.Block):
-#         """A miscellaneous block."""
-
-#         NAME = 'misc'
-#         OPTIONS = {'test': ['whatever', block.type_string]}
-
-#         def on_create(self, parent):
-#             """Create."""
-
-#             return etree.SubElement(parent, 'div', {'test': self.options['test']})
-
-#     extension = ['pymdownx.blocks', 'pymdownx.superfences']
-#     extension_configs = {'pymdownx.blocks': {'blocks': {MiscBlock: None}}}
-
-#     def test_no_config(self):
-#         """Test no YAML config and no new line."""
-
-#         self.check_markdown(
-#             R'''
-#             /// misc
-#             content
-#             ///
-#             ''',
-#             '''
-#             <div test="whatever">
-#             <p>content</p>
-#             </div>
-#             ''',
-#             True
-#         )
-
-#     def test_config_no_new_line(self):
-#         """Test YAML config and no new line."""
-
-#         self.check_markdown(
-#             R'''
-#             /// misc
-#                 test: tag
-#             content
-#             ///
-#             ''',
-#             '''
-#             <div test="tag">
-#             <p>content</p>
-#             </div>
-#             ''',
-#             True
-#         )
-
-#     def test_config_no_fence(self):
-#         """Test YAML config and no fence."""
-
-#         self.check_markdown(
-#             R'''
-#             /// misc
-#             test: tag
-
-#             content
-#             ///
-#             ''',
-#             '''
-#             <div test="whatever">
-#             <p>test: tag</p>
-#             <p>content</p>
-#             </div>
-#             ''',
-#             True
-#         )
-
-#     def test_config_with_new_line_before(self):
-#         """Test YAML config and with new line before it."""
-
-#         self.check_markdown(
-#             R'''
-#             /// misc
-
-#                 test: tag
-
-#             content
-#             ///
-#             ''',
-#             '''
-#             <div test="whatever">
-#             <pre><code>test: tag
-#             </code></pre>
-#             <p>content</p>
-#             </div>
-#             ''',
-#             True
-#         )
-
-
-# class TestBadArgOptionParsers(util.MdCase):
-#     """Test when a block's options do not fulfill the parser's expectations."""
-
-#     class FailBlock(block.Block):
-#         """Test failure to satisfy parser."""
-
-#         NAME = 'fail'
-#         ARGUMENTS = {'required': 1, 'parsers': [block.type_html_identifier]}
-#         OPTIONS = {'test': ['tag', block.type_html_identifier]}
-
-#         def on_create(self, parent):
-#             """Create."""
-
-#             return etree.SubElement(parent, 'div')
-
-#     extension = ['pymdownx.blocks']
-#     extension_configs = {'pymdownx.blocks': {'blocks': {FailBlock: None}}}
-
-#     def test_fail_args(self):
-#         """Test failure of arguments."""
-
-#         self.check_markdown(
-#             R'''
-#             /// fail | 3tag
-#             ///
-#             ''',
-#             '''
-#             <p>/// fail | 3tag
-#             ///</p>
-#             ''',
-#             True
-#         )
-
-#     def test_fail_opts(self):
-#         """Test failure of options."""
-
-#         self.check_markdown(
-#             R'''
-#             /// fail | tag
-#                 test: 3tag
-
-#             content
-#             ///
-#             ''',
-#             '''
-#             <p>/// fail | tag
-#                 test: 3tag</p>
-#             <p>content
-#             ///</p>
-#             ''',
-#             True
-#         )
-
-#     def test_bad_config(self):
-#         """Test when content is used in place of config."""
-
-#         self.check_markdown(
-#             R'''
-#             /// fail | tag
-#                 content
-#             ///
-#             ''',
-#             '''
-#             <p>/// fail | tag
-#                 content
-#             ///</p>
-#             ''',
-#             True
-#         )
-
-#     def test_bad_yaml(self):
-#         """Test config is bad YAML."""
-
-#         self.check_markdown(
-#             R'''
-#             /// fail | tag
-#                 :
-#             ///
-#             ''',
-#             '''
-#             <p>/// fail | tag
-#                 :
-#             ///</p>
-#             ''',
-#             True
-#         )
-
-
-# class TestBlockSplit(util.MdCase):
-#     """Test Blocks that split arguments."""
-
-#     class SplitBlock(block.Block):
-#         """Split block."""
-
-#         NAME = 'split'
-
-#         ARGUMENTS = {'required': 2}
-
-#         def on_create(self, parent):
-#             """Create."""
-
-#             return etree.SubElement(parent, 'div', {self.args[0]: '0', self.args[1]: '1'})
-
-#     extension = ['pymdownx.blocks']
-#     extension_configs = {'pymdownx.blocks': {'blocks': {SplitBlock: None}}}
-
-#     def test_split(self):
-#         """Test that attributes cannot be overridden."""
-
-#         self.check_markdown(
-#             R'''
-#             /// split | this that
-#             ///
-#             ''',
-#             '''
-#             <div that="1" this="0"></div>
-#             ''',
-#             True
-#         )
-
-#     def test_split_less(self):
-#         """Test failure when arguments are too few."""
-
-#         self.check_markdown(
-#             R'''
-#             /// split | this
-#             ///
-#             ''',
-#             '''
-#             <p>/// split | this
-#             ///</p>
-#             ''',
-#             True
-#         )
-
-#     def test_split_greater(self):
-#         """Test failure when arguments are too many."""
-
-#         self.check_markdown(
-#             R'''
-#             /// split | this that another
-#             ///
-#             ''',
-#             '''
-#             <p>/// split | this that another
-#             ///</p>
-#             ''',
-#             True
-#         )
+class TestRegister(unittest.TestCase):
+    """Test registration cases."""
+
+    def test_duplicates(self):
+        """Test duplicates."""
+
+        with self.assertRaises(ValueError):
+            markdown.markdown('test', extensions=['pymdownx.blocks.admonition', 'pymdownx.blocks.admonition'])
+
+
+class TestBlockUndefinedOption(util.MdCase):
+    """Test Blocks with undefined options."""
+
+    extension = ['pymdownx.blocks.html', 'pymdownx.blocks.definition']
+
+    def test_undefined_option(self):
+        """An undefined option will cause the block parsing to fail."""
+
+        self.check_markdown(
+            R'''
+            /// html | div
+                option: whatever
+
+            content
+            ///
+            ''',
+            '''
+            <p>/// html | div
+                option: whatever</p>
+            <p>content
+            ///</p>
+            ''',
+            True
+        )
+
+    def test_bad_option(self):
+        """An undefined option will cause the block parsing to fail."""
+
+        self.check_markdown(
+            R'''
+            /// html | div
+                attrs: whatever
+
+            content
+            ///
+            ''',
+            '''
+            <p>/// html | div
+                attrs: whatever</p>
+            <p>content
+            ///</p>
+            ''',
+            True
+        )
+
+    def test_no_arg(self):
+        """Test no options."""
+
+        self.check_markdown(
+            R'''
+            /// html
+                attrs: whatever
+
+            content
+            ///
+            ''',
+            '''
+            <p>/// html
+                attrs: whatever</p>
+            <p>content
+            ///</p>
+            ''',
+            True
+        )
+
+    def test_too_many_args(self):
+        """Test too many options."""
+
+        self.check_markdown(
+            R'''
+            /// define
+                option: whatever
+
+            content
+            ///
+            ''',
+            '''
+            <p>/// define
+                option: whatever</p>
+            <p>content
+            ///</p>
+            ''',
+            True
+        )
+
+    def test_bad_frontmatter(self):
+        """Test bad frontmatter."""
+
+        self.check_markdown(
+            R'''
+            /// define
+                bad()
+
+            content
+            ///
+            ''',
+            '''
+            <p>/// define
+                bad()</p>
+            <p>content
+            ///</p>
+            ''',
+            True
+        )
+
+    def test_bad_frontmatter2(self):
+        """Test bad frontmatter."""
+
+        self.check_markdown(
+            R'''
+            /// define
+                &
+
+            content
+            ///
+            ''',
+            '''
+            <p>/// define
+                &amp;</p>
+            <p>content
+            ///</p>
+            ''',
+            True
+        )
+
+
+class TestCodeHandling(util.MdCase):
+    """Test some code related cases."""
+
+    extension = ['pymdownx.blocks.html', 'pymdownx.superfences']
+
+    def test_superfence_block(self):
+        """Test blocks with fenced code content."""
+
+        self.check_markdown(
+            R'''
+            /// html | div
+            ```python
+            import foo
+            ```
+            ///
+            ''',
+            '''
+            <div>
+            <div class="highlight"><pre><span></span><code><span class="kn">import</span> <span class="nn">foo</span>
+            </code></pre></div>
+            </div>
+            ''',
+            True
+        )
+
+    def test_superfence_inline(self):
+        """Test blocks with fenced code content."""
+
+        self.check_markdown(
+            R'''
+            /// html | span
+            ```python
+            import foo
+            ```
+
+            Other content
+            ///
+            ''',
+            '''
+            <span><code>python
+            import foo</code>
+
+            Other content</span>
+            ''',
+            True
+        )
 
 
 class TestAttributes(util.MdCase):